Refactor auth implementation
