+++ src/routes/auth.ts
@@ -1,8 +1,13 @@
 import { Router, type Request, type Response } from "express";
-import { newUserSchema, users, type NewUser } from "../db/schema";
+import { newUserSchema, users, type NewUser, type User } from "../db/schema";
+import { loginSchema } from "../utils/definitions";
 import { db } from "../db/db";
 
-const bcrypt = require("bcrypt");
+import { eq } from "drizzle-orm";
+import { ZodError } from "zod";
+import * as jwt from "jsonwebtoken";
+import * as bcrypt from "bcrypt";
+
 const saltRounds = 10;
 
 export const authRouter: Router = Router();
@@ -14,7 +19,7 @@ authRouter.post("/signup", async (req: Request, res: Response) => {
   if (!parsedUser.success) {
     return res.status(400).send("Invalid request");
   }
-  console.log("Reached!");
+
   try {
     let userData = parsedUser!.data as NewUser;
     // Hash password for storage
@@ -28,5 +33,46 @@ authRouter.post("/signup", async (req: Request, res: Response) => {
     return res.status(200).send(`Inserted new user ${parsedUser.data?.email}`);
   } catch (error) {
     console.log(error);
+  }
+});
+
+authRouter.post("/login", async (req: Request, res: Response) => {
+  try {
+    const result = loginSchema.safeParse(req.body);
+    if (!result.success) {
+      const err: ZodError = result.error;
+      return res.status(400).json({ errors: err.errors });
+    }
+
+    const { numeroCuenta, password } = req.body;
+
+    // agregar que result es de tipo user
+    const userQuery: User[] | any = await db
+      .select()
+      .from(users)
+      .where(eq(users.numeroCuenta, numeroCuenta));
+
+    const user = userQuery[0];
+    if (!user)
+      return res.status(401).json({ msg: "User credentials are not valid" });
+
+    const matchPassword = await bcrypt.compare(password, user.password);
+
+    if (!matchPassword)
+      return res.status(401).json({ msg: "User credentials are not valid" });
+
+    // creating cookie session
+    const SECRET_KEY: string = process.env.SECRET_KEY || "prueba";
+    if (!SECRET_KEY) {
+      return res.status(500).json({ msg: "Could not authenticate user" });
+    }
+
+    const token = jwt.sign({ userId: user.numeroCuenta }, SECRET_KEY, {
+      expiresIn: "7d",
+    });
+
+    res.status(200).cookie("access_token", token).json(user);
+  } catch (error) {
+    return res.status(500).json({ msg: "Error while login", e: error });
   }
 });
